global refinements 05.04 - removed old contracts, containers & docs dirs, .reduxrc, DEPLOY.md, unnecessary components/common/index.js - formatting.js moved to utils dir - PromptPassword file naming style fix - other minor fixes
