--- .reduxrc
@@ -1,7 +1,0 @@
-{
-    "sourceBase": "src",
-    "testBase":"test",
-    "smartPath":"containers",
-    "dumbPath":"components",
-    "fileCasing":"default"
-}
