--- old/ConvertLib.sol
@@ -1,8 +1,0 @@
-pragma solidity ^0.4.2;
-
-library ConvertLib{
-	function convert(uint amount,uint conversionRate) returns (uint convertedAmount)
-	{
-		return amount * conversionRate;
-	}
-}
