--- old/ERC20Interface.sol
@@ -1,13 +1,0 @@
-pragma solidity ^0.4.4;
-
-contract ERC20Interface {
-    event Transfer(address indexed _from, address indexed _to, uint256 _value);
-    event Approved(address indexed _owner, address indexed _spender, uint256 _value);
-
-    function totalSupply() constant returns (uint256 supply);
-    function balanceOf(address _owner) constant returns (uint256 balance);
-    function transfer(address _to, uint256 _value) returns (bool success);
-    function transferFrom(address _from, address _to, uint256 _value) returns (bool success);
-    function approve(address _spender, uint256 _value) returns (bool success);
-    function allowance(address _owner, address _spender) constant returns (uint256 remaining);
-}
