--- old/Stub.sol
@@ -1,5 +1,0 @@
-pragma solidity ^0.4.4;
-
-contract Stub {
-    function () {}
-}
