+++ package.json
@@ -28,7 +28,6 @@
     "chai-as-promised": "^6.0.0",
     "chai-enzyme": "^0.6.1",
     "chalk": "1.1.3",
-    "classnames": "^2.2.5",
     "core-decorators": ">=0.12",
     "coveralls": "^2.12.0",
     "cross-spawn": "4.0.0",
@@ -115,6 +114,7 @@
     "ipfs": "0.22.0",
     "ipfs-repo": "0.11.3",
     "moment": "^2.17.1",
+    "numeral": "^2.0.6",
     "orbit-db": "^0.16.3",
     "orbit-db-pubsub": "^0.1.6",
     "react": "^15.2.1",
