+++ src/components/modals/Modal.js
@@ -20,24 +20,24 @@ import {
   SETTINGS_OTHER_CONTRACT_TYPE,
   SETTINGS_OTHER_CONTRACT_MODIFY_TYPE,
   IPFS_TYPE
-} from '../redux/ui/modal.js'
-import PromptPassword from '../components/modals/prompt_password'
-import AlertModal from '../components/modals/AlertModal'
-import LOCModal from '../components/modals/LOCModal'
-import SendToExchangeModal from '../components/modals/SendToExchangeModal'
-import ChangeNumberSignaturesModal from '../components/modals/ChangeNumberSignaturesModal'
-import IssueLHForm from '../components/modals/IssueLHModal'
-import RedeemLHModal from '../components/modals/RedeemLHModal'
-import UploadedFileModal from '../components/modals/UploadedFileModal'
-import NewPollModal from '../components/modals/NewPollModal'
-import PollModal from '../components/modals/poll/PollModal'
-import DepositTimeModal from '../components/modals/DepositTimeModal'
-import SettingsCBEModal from '../components/modals/settings/CBEAddressModal'
-import SettingsTokenViewModal from '../components/modals/settings/TokenViewModal'
-import SettingsTokenModal from '../components/modals/settings/TokenModal'
-import SettingsOtherContractModal from '../components/modals/settings/OtherContractModal'
-import SettingsOtherContractModifyModal from '../components/modals/settings/OtherContractModifyModal'
-import IPFSFileUpload from '../components/modals/IPFSFileUpload'
+} from '../../redux/ui/modal.js'
+import PromptPassword from './PromptPassword'
+import AlertModal from './AlertModal'
+import LOCModal from './LOCModal'
+import SendToExchangeModal from './SendToExchangeModal'
+import ChangeNumberSignaturesModal from './ChangeNumberSignaturesModal'
+import IssueLHForm from './IssueLHModal'
+import RedeemLHModal from './RedeemLHModal'
+import UploadedFileModal from './UploadedFileModal'
+import NewPollModal from './NewPollModal'
+import PollModal from './poll/PollModal'
+import DepositTimeModal from './DepositTimeModal'
+import SettingsCBEModal from './settings/CBEAddressModal'
+import SettingsTokenViewModal from './settings/TokenViewModal'
+import SettingsTokenModal from './settings/TokenModal'
+import SettingsOtherContractModal from './settings/OtherContractModal'
+import SettingsOtherContractModifyModal from './settings/OtherContractModifyModal'
+import IPFSFileUpload from './IPFSFileUpload'
 
 const mapDispatchToProps = (dispatch) => ({
   hideModal: () => dispatch(hideModal())
